Solution to Convert Decimal to Binary

+++ Convert Decimal to Binary@@67882a731b4392d8fc1577f5/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 
 int main(){
-    int n;
+    long int n;
     scanf("%d", &n);
     int rem, ans = 0, pow = 1;
     while (n > 0){
